services/api_gateway/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
asyncpg==0.29.0
alembic==1.13.3
redis==5.1.1
python-jose[cryptography]==3.3.0
bcrypt==4.2.1
python-multipart==0.0.12
httpx==0.27.2
pydantic[email]==2.10.6
pydantic-settings==2.7.0
cryptography==44.0.2
livekit-api==0.7.1
sendgrid==6.11.0
python-dotenv==1.0.1
